Update directive to render generic tags - #3984
