+++ ui/craftercms-sphinx-theme/src/static/scss/_custom.scss
@@ -336,32 +336,35 @@ div[class^="highlight-"] {
   }
 }
 
-.versions_tags {
+.version-tag {
+  color: white;
+  display: inline-block;
+  background: #595959;
+  border-radius: 5px;
+  font-size: 15px;
+
+  &.inline {
+    margin-bottom: 10px;
+  }
+
+  .version__label {
+    margin-right: 0;
+    display: inline-block;
+    padding: 2px 10px;
+  }
+
+  .version__number {
+    background-color: #2284bb;
+    margin-left: 0;
+    display: inline-block;
+    padding: 2px 10px;
+    border-radius: 0 5px 5px 0;
+  }
+}
+
+.versions-tags {
   &.inline {
     display: inline-block;
     margin-bottom: 10px;
-    float: right;
-  }
-
-  .version {
-    color: white;
-    display: inline-block;
-    background: #595959;
-    border-radius: 5px;
-    font-size: 15px;
-
-    .version__label {
-      margin-right: 0;
-      display: inline-block;
-      padding: 2px 10px;
-    }
-
-    .version__number {
-      background-color: #2284bb;
-      margin-left: 0;
-      display: inline-block;
-      padding: 2px 10px;
-      border-radius: 0 5px 5px 0;
-    }
   }
 }
